todolist_core/requirements.txt
amqp==5.3.1
asciimatics==1.9.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.2
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt== 5.3.1
drf-yasg==1.21.8
future==1.0.0
inflection==0.5.1
kombu==5.4.2
Markdown==3.7
packaging==24.1
pillow==11.0.0
pip==24.3.1
prompt_toolkit==3.0.48
pyfiglet==1.0.2
PyJWT==2.9.0
pypiwin32==223
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
redis==5.2.0
six==1.16.0
sqlparse==0.5.1
todolist==0.6.0
tzdata==2024.2
uritemplate==4.1.1
views==0.3
vine==5.1.0
wcwidth==0.2.13